+++ NiBAx/plugins/loadsave/loadsave.ui
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ui version="4.0">
- <class>Data</class>
- <widget class="QWidget" name="Data">
+ <class>LoadSave</class>
+ <widget class="QWidget" name="LoadSave">
   <property name="geometry">
    <rect>
     <x>0</x>
